parity/requirements.txt
pandas
numpy
ruamel.yaml